requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
Faker==2.0.1
importlib-metadata==0.19
more-itertools==7.2.0
packaging==19.1
Pillow==6.1.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
pytest==5.1.1
pytest-html==1.22.0
pytest-metadata==1.8.0
python-dateutil==2.8.0
selenium==3.141.0
six==1.12.0
text-unidecode==1.2
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.6.0
